more tweaks
